Синди туррели теперь обращают внимание на карты агента + Разные типы крови для Тайпановцев
* Синди туррели теперь обращают внимание на карту агента (Так как использование только прописанной в самих мобах фракции не очень надёжно. Например она слетает при клонировании и внезапно свои туррели начинают агрессировать на вас.)
* Теперь агентам при спавне выдаётся рандомный тип крови. До этого у всех было "A+"

+++ code/game/machinery/portable_turret.dm
@@ -582,6 +582,10 @@ GLOBAL_LIST_EMPTY(turret_icons)
 	if(in_faction(L))
 		return TURRET_NOT_TARGET
 
+	var/obj/item/card/CID = L.get_id_card()
+	if(istype(CID, /obj/item/card/id/syndicate) && src.faction == "syndicate")
+		return TURRET_NOT_TARGET
+
 	if(lethal && locate(/mob/living/silicon/ai) in get_turf(L))		//don't accidentally kill the AI!
 		return TURRET_NOT_TARGET
